Fix dev-dependencies and test files requires
React was setup before Jest, so tests were never executed when I
initially ported the project to npm and wally. React could not even
depend on the jest packages.

This PR is a first step towards making the test suite run: it corrects
a lot of the requires in test files and also define each dev-dependency.

This PR also enables the `@jsdotlua/react-noop-renderer` and
`@jsdotlua/jest-react` to now be published!